\name{wgcna_corr-package}
\alias{wgcna_corr-package}
\alias{wgcna_corr}
\docType{package}
\title{
\packageTitle{wgcna_corr}
}
\description{
\packageDescription{wgcna_corr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{wgcna_corr}
\packageIndices{wgcna_corr}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{wgcna_corr}

Maintainer: \packageMaintainer{wgcna_corr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
